+++ cas_server/tests/test_federate.py
@@ -8,16 +8,16 @@
 # along with this program; if not, write to the Free Software Foundation, Inc., 51
 # Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #
-# (c) 2016 Valentin Samir
+# (c) 2016-2025 Valentin Samir
 """tests for the CAS federate mode"""
+import importlib
+
 from cas_server import default_settings
 from cas_server.default_settings import settings
 
 import django
 from django.test import TestCase, Client
 from django.test.utils import override_settings
-
-from six.moves import reload_module
 
 from cas_server import utils, models
 from cas_server.tests.mixin import BaseServicePattern, CanLogin, FederatedIendityProviderModel
@@ -49,7 +49,7 @@ class FederateAuthLoginLogoutTestCase(
     def test_default_settings(self):
         """default settings should populated some default variable then CAS_FEDERATE is True"""
         del settings.CAS_AUTH_CLASS
-        reload_module(default_settings)
+        importlib.reload(default_settings)
         self.assertEqual(settings.CAS_AUTH_CLASS, "cas_server.auth.CASFederateAuth")
 
     def test_login_get_provider(self):
@@ -64,7 +64,7 @@ class FederateAuthLoginLogoutTestCase(
             ) in response.content.decode("utf-8"))
         self.assertEqual(response.context['post_url'], '/federate')
 
-    def test_login_post_provider(self, remember=False):
+    def test_login_post_provider(self, remember=False, ret=False):
         """test a successful login wrokflow"""
         tickets = []
         # choose the example.com provider
@@ -157,12 +157,13 @@ class FederateAuthLoginLogoutTestCase(
             response = client.get("/login", {'service': self.service})
             self.assertEqual(response.status_code, 302)
             self.assertTrue(response["Location"].startswith("%s?ticket=" % self.service))
-        return tickets
+        if ret:
+            return tickets
 
     def test_login_twice(self):
         """Test that user id db is used for the second login (cf coverage)"""
         self.test_login_post_provider()
-        tickets = self.test_login_post_provider()
+        tickets = self.test_login_post_provider(ret=True)
         # trying to authenticated while being already authenticated should redirect to /login
         for (provider, _, client) in tickets:
             response = client.get("/federate/%s" % provider.suffix)
@@ -256,7 +257,7 @@ class FederateAuthLoginLogoutTestCase(
     def test_auth_federate_slo(self):
         """test that SLO receive from backend CAS log out the users"""
         # get tickets and connected clients
-        tickets = self.test_login_post_provider()
+        tickets = self.test_login_post_provider(ret=True)
         for (provider, ticket, client) in tickets:
             # SLO for an unkown ticket should do nothing
             response = client.post(
@@ -307,7 +308,7 @@ class FederateAuthLoginLogoutTestCase(
             and redirected to his CAS logout page
         """
         # get tickets and connected clients, then follow normal logout
-        tickets = self.test_login_post_provider()
+        tickets = self.test_login_post_provider(ret=True)
         for (provider, _, client) in tickets:
             response = client.get("/logout")
             self.assertEqual(response.status_code, 302)
@@ -328,7 +329,7 @@ class FederateAuthLoginLogoutTestCase(
                 ) in response.content
             )
 
-        tickets = self.test_login_post_provider()
+        tickets = self.test_login_post_provider(ret=True)
         if django.VERSION >= (1, 8):
             # assume the username session variable has been tempered (should not happend)
             for (provider, _, client) in tickets:
@@ -345,7 +346,7 @@ class FederateAuthLoginLogoutTestCase(
             If the user check remember, next login should not offer the chose of the backend CAS
             and use the one store in the cookie
         """
-        tickets = self.test_login_post_provider(remember=True)
+        tickets = self.test_login_post_provider(remember=True, ret=True)
         for (provider, _, client) in tickets:
             client.get("/logout")
             response = client.get("/login")
@@ -357,7 +358,7 @@ class FederateAuthLoginLogoutTestCase(
 
     def test_forget_provider(self):
         """Test the logout option to forget remembered provider"""
-        tickets = self.test_login_post_provider(remember=True)
+        tickets = self.test_login_post_provider(remember=True, ret=True)
         for (provider, _, client) in tickets:
             self.assertIn("remember_provider", client.cookies)
             self.assertEqual(client.cookies["remember_provider"].value, provider.suffix)
@@ -369,7 +370,7 @@ class FederateAuthLoginLogoutTestCase(
         """
             Test authentication renewal with federation mode
         """
-        tickets = self.test_login_post_provider()
+        tickets = self.test_login_post_provider(ret=True)
         for (provider, _, client) in tickets:
             # Try to renew authentication(client already authenticated in test_login_post_provider
             response = client.get("/login?renew=true")