+++ cas_server/tests/utils.py
@@ -8,7 +8,7 @@
 # along with this program; if not, write to the Free Software Foundation, Inc., 51
 # Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #
-# (c) 2016 Valentin Samir
+# (c) 2016-2025 Valentin Samir
 """Some utils functions for tests"""
 from cas_server.default_settings import settings
 
@@ -17,12 +17,10 @@ from django.test import Client
 from django.template import loader
 from django.utils import timezone
 
-import cgi
-import six
 from threading import Thread
 from lxml import etree
-from six.moves import BaseHTTPServer
-from six.moves.urllib.parse import urlparse, parse_qsl, parse_qs
+import http.server as BaseHTTPServer
+from urllib.parse import urlparse, parse_qsl, parse_qs
 from datetime import timedelta
 
 from cas_server import models
@@ -43,7 +41,7 @@ else:
 
 def return_unicode(string, charset):
     """make `string` a unicode if `string` is a unicode or bytes encoded with `charset`"""
-    if not isinstance(string, six.text_type):
+    if not isinstance(string, str):
         return string.decode(charset)
     else:
         return string
@@ -54,7 +52,7 @@ def return_bytes(string, charset):
         make `string` a bytes encoded with `charset` if `string` is a unicode
         or bytes encoded with `charset`
     """
-    if isinstance(string, six.text_type):
+    if isinstance(string, str):
         return string.encode(charset)
     else:
         return string
@@ -156,6 +154,40 @@ def get_proxy_ticket(service):
     return ticket
 
 
+def cgi_parse_header(line):
+    """Parse a Content-type like header.
+
+    Return the main content-type and a dictionary of options.
+
+    """
+
+    def _parseparam(s):
+        while s[:1] == ';':
+            s = s[1:]
+            end = s.find(';')
+            while end > 0 and (s.count('"', 0, end) - s.count('\\"', 0, end)) % 2:
+                end = s.find(';', end + 1)
+            if end < 0:
+                end = len(s)
+            f = s[:end]
+            yield f.strip()
+            s = s[end:]
+
+    parts = _parseparam(';' + line)
+    key = parts.__next__()
+    pdict = {}
+    for p in parts:
+        i = p.find('=')
+        if i >= 0:
+            name = p[:i].strip().lower()
+            value = p[i+1:].strip()
+            if len(value) >= 2 and value[0] == value[-1] == '"':
+                value = value[1:-1]
+                value = value.replace('\\\\', '\\').replace('\\"', '"')
+            pdict[name] = value
+    return key, pdict
+
+
 class HttpParamsHandler(BaseHTTPServer.BaseHTTPRequestHandler):
     """
         A simple http server that return 200 on GET or POST
@@ -174,9 +206,9 @@ class HttpParamsHandler(BaseHTTPServer.BaseHTTPRequestHandler):
 
     def do_POST(self):
         """Called on a POST request on the BaseHTTPServer"""
-        ctype, pdict = cgi.parse_header(self.headers.get('content-type'))
+        ctype, pdict = cgi_parse_header(self.headers.get('content-type'))
         if ctype == 'multipart/form-data':
-            postvars = cgi.parse_multipart(self.rfile, pdict)
+            raise NotImplementedError()
         elif ctype == 'application/x-www-form-urlencoded':
             length = int(self.headers.get('content-length'))
             postvars = parse_qs(self.rfile.read(length), keep_blank_values=1)